requirements.txt
#Flask==0.10.1
#Flask-Script==2.0.5
#flask-mongoengine==0.10.6
#flask-wtf(FlaskForm)
Flask-WTF==0.13.1
py-bcrypt==0.4
#signal-mongoengine
blinker==1.4
boto3==1.3.0
#sudo apt-get install imagemagick libmagickcore-dev
#sudo apt-get update
#sudo apt-get install libmagickcore-dev
#handler image
wand==0.4.2
#time
Arrow==0.8.0
#html clearner
bleach==1.4.3

############################################################
numpy            ==         1.11.2             
openpyxl        ==          2.4.1          
pandas          ==          0.19.2                    
requests==                  2.12.4                  
xlrd==                      1.0.0               
xlsxwriter==                0.9.6                   
xlwt==                      1.2.0 
bokeh== 0.12.4
#docker-compose run web pip install pandas
#Find the docker web container name and run: docker exec -it petsapi_web_1 python tests.py
#Find the docker web container name and run: docker exec -it petsapi_db_1 mongo
#Checkout code on /opt/pets-api on local computer
#Build the containers: docker-compose build
#Copy settings.py.bak to settings.py and add 'mongodb' as MONGODB_HOST
#Start mongodb separately: docker-compose up -d db
#Start the app: docker-compose up web
#To start the application with pdb enabled: docker-compose run --service-ports web
############################################################
pymongo==2.8
mongoengine==0.8
Flask==0.11
Flask-Script==2.0.5
flask-mongoengine==0.8.2
py-bcrypt==0.4
# pip install -U mongoengine
